pyproject.toml
[project]
name = "real-time-object-detection-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "streamlit>=1.53.1",
    "ultralytics>=8.4.7",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]

[tool.setuptools]
packages = ["api", "ui"]

[tool.pytest.ini_options]
markers = ["e2e: End-to-end tests with real model (slower, no mocks)"]
